refactor(): 수익률 식, 제네릭 타입 선언 수정

+++ src/main/java/lotto/Buyer.java
@@ -30,7 +30,7 @@ public class Buyer {
         ResultView.printResult(this, lottoGame);
     }
 
-    public double getPurchaseAmount() {
+    public int getPurchaseAmount() {
         return purchaseAmount;
     }
 }
